+++ lib/240313/Intangible_Asset.dart
@@ -1,0 +1,12 @@
+//무형자산 클래스
+import 'package:learn_dart_together/240313/asset.dart';
+
+abstract class IntangibleAsset extends Asset {
+  // String name;
+  // int price;
+  // String color;
+
+  IntangibleAsset(super.name, super.price);
+}
+
+
